Fix truncation warning with explicit cast

+++ src/core/hle/D3D8/Direct3D9/WalkIndexBuffer.cpp
@@ -58,7 +58,7 @@ void WalkIndexBuffer_SSE41(XTL::INDEX16 & LowIndex, XTL::INDEX16 & HighIndex, XT
 	}
 
 	__m128i *unalignedIndices = (__m128i*) pIndexData;\
-	__m128i min = _mm_set1_epi16(USHRT_MAX);
+	__m128i min = _mm_set1_epi16((short)USHRT_MAX);
 	__m128i max = _mm_setzero_si128();
 
 	// Min / max over index data
@@ -72,7 +72,7 @@ void WalkIndexBuffer_SSE41(XTL::INDEX16 & LowIndex, XTL::INDEX16 & HighIndex, XT
 	min = _mm_minpos_epu16(min);
 
 	// horizontal max (using minpos)
-	max = _mm_subs_epu16(_mm_set1_epi16(USHRT_MAX), max); //invert
+	max = _mm_subs_epu16(_mm_set1_epi16((short)USHRT_MAX), max); //invert
 	max = _mm_minpos_epu16(max);
 
 	// Get the min and max out
